sentiment analysis of email collection

+++ src/Visualizable.java
@@ -1,8 +1,8 @@
 // Any class that implements this interace is guaranteed to be able to
 // represented as an L-System
 public interface Visualizable {
-  int sentiment();
+  float sentiment();
   int age();
-  int average_email_length();
+  float average_email_length();
 }
